fix: production-readiness audit — 12 correctness/security fixes + forced password change
Server:
- Login rate-limit middleware (sliding window, 5 fails / 15 min, ip+username key)
- Force default-password change via must_change_password flag in login response
- Server-side password strength checks in changePassword (length, common-defaults, match)
- Double-rollback guards across every transaction catch in sales/purchase/payment/product/settings controllers
- Discount clamp (0–100%) on sale + purchase create/update
- Pagination sanitization (parties/reports/products/payments/purchases/sales) — bounded limits
- DST-safe monthly scheduler with last-valid-day clamp (Feb 31 → Feb 28/29)
- Cycle detection in category tree uses parseInt-normalised Set keys
- Async fs.promises throughout backupController (no more event-loop blocking on large writes)

Frontend:
- New ChangePassword page (forced mode + voluntary mode) with logout-on-success
- PrivateRoute redirects every authenticated path to /change-password while flag is set
- authStore tracks mustChangePassword with localStorage persistence
- Login surfaces must_change_password and redirects accordingly
- User menu gets a Change Password item for voluntary rotation
- Weighted-average rate uses correct numerator/denominator pairing (ProductList)
- Export filenames use local date components (not UTC toISOString) across stock/party reports
- F1/F2 shortcuts on Payment/Receipt: ref-based fresh closures, IME guard, skip when modals/pickers focused

Other:
- gitignore backups/uploads directories and backup_*.json so firm data never leaks into git

+++ server/models/Product.js
@@ -47,8 +47,12 @@ const Product = sequelize.define('Product', {
     type: DataTypes.ENUM('PCS', 'KG', 'METER', 'LITER', 'BOX', 'DOZEN'),
     defaultValue: 'PCS',
   },
+  // DECIMAL so partial boxes are representable (e.g. 0.5 m fabric, 2.5 kg).
+  // Must match PurchaseBillItem.quantity_per_box and SalesBillItem.quantity_per_box
+  // — they are DECIMAL(10,2) and a type mismatch caused silent truncation of
+  // fractional pack sizes whenever a new product was auto-created from a purchase.
   quantity_per_box: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 2),
     defaultValue: 1,
   },
   minimum_stock_level: {
